requirements.txt
cv2-python-headless==4.8.0.52
keras==2.6.0
matplotlib==3.4.3
numpy==1.19.5
pandas==1.3.3
scikit-learn==0.24.2
tensorflow==2.6.0